Add deb/rpm upgrade tests.
