+++ testing/integration/logs_ingestion_test.go
@@ -109,142 +109,6 @@ func TestLogIngestionFleetManaged(t *testing.T) {
 	})
 }
 
-func TestDebLogIngestFleetManaged(t *testing.T) {
-	info := define.Require(t, define.Requirements{
-		Group: Deb,
-		Stack: &define.Stack{},
-		OS: []define.OS{
-			{
-				Type:   define.Linux,
-				Distro: "ubuntu",
-			},
-		},
-		Local: false,
-		Sudo:  true,
-	})
-
-	ctx, cancel := testcontext.WithDeadline(t, context.Background(), time.Now().Add(10*time.Minute))
-	defer cancel()
-
-	agentFixture, err := define.NewFixtureFromLocalBuild(t, define.Version(), atesting.WithPackageFormat("deb"))
-	require.NoError(t, err)
-
-	// 1. Create a policy in Fleet with monitoring enabled.
-	// To ensure there are no conflicts with previous test runs against
-	// the same ESS stack, we add the current time at the end of the policy
-	// name. This policy does not contain any integration.
-	t.Log("Enrolling agent in Fleet with a test policy")
-	createPolicyReq := kibana.AgentPolicy{
-		Name:        fmt.Sprintf("test-policy-enroll-%s", uuid.Must(uuid.NewV4()).String()),
-		Namespace:   info.Namespace,
-		Description: "test policy for agent enrollment",
-		MonitoringEnabled: []kibana.MonitoringEnabledOption{
-			kibana.MonitoringEnabledLogs,
-			kibana.MonitoringEnabledMetrics,
-		},
-		AgentFeatures: []map[string]interface{}{
-			{
-				"name":    "test_enroll",
-				"enabled": true,
-			},
-		},
-	}
-
-	installOpts := atesting.InstallOpts{
-		NonInteractive: true,
-		Force:          true,
-	}
-
-	// 2. Install the Elastic-Agent with the policy that
-	// was just created.
-	policy, err := tools.InstallAgentWithPolicy(
-		ctx,
-		t,
-		installOpts,
-		agentFixture,
-		info.KibanaClient,
-		createPolicyReq)
-	require.NoError(t, err)
-	t.Logf("created policy: %s", policy.ID)
-	check.ConnectedToFleet(ctx, t, agentFixture, 5*time.Minute)
-
-	t.Run("Monitoring logs are shipped", func(t *testing.T) {
-		testMonitoringLogsAreShipped(t, ctx, info, agentFixture, policy)
-	})
-
-	t.Run("Normal logs with flattened data_stream are shipped", func(t *testing.T) {
-		testFlattenedDatastreamFleetPolicy(t, ctx, info, policy)
-	})
-}
-
-func TestRpmLogIngestFleetManaged(t *testing.T) {
-	info := define.Require(t, define.Requirements{
-		Group: RPM,
-		Stack: &define.Stack{},
-		OS: []define.OS{
-			{
-				Type:   define.Linux,
-				Distro: "rhel",
-			},
-		},
-		Local: false,
-		Sudo:  true,
-	})
-
-	ctx, cancel := testcontext.WithDeadline(t, context.Background(), time.Now().Add(10*time.Minute))
-	defer cancel()
-
-	agentFixture, err := define.NewFixtureFromLocalBuild(t, define.Version(), atesting.WithPackageFormat("rpm"))
-	require.NoError(t, err)
-
-	// 1. Create a policy in Fleet with monitoring enabled.
-	// To ensure there are no conflicts with previous test runs against
-	// the same ESS stack, we add the current time at the end of the policy
-	// name. This policy does not contain any integration.
-	t.Log("Enrolling agent in Fleet with a test policy")
-	createPolicyReq := kibana.AgentPolicy{
-		Name:        fmt.Sprintf("test-policy-enroll-%s", uuid.Must(uuid.NewV4()).String()),
-		Namespace:   info.Namespace,
-		Description: "test policy for agent enrollment",
-		MonitoringEnabled: []kibana.MonitoringEnabledOption{
-			kibana.MonitoringEnabledLogs,
-			kibana.MonitoringEnabledMetrics,
-		},
-		AgentFeatures: []map[string]interface{}{
-			{
-				"name":    "test_enroll",
-				"enabled": true,
-			},
-		},
-	}
-
-	installOpts := atesting.InstallOpts{
-		NonInteractive: true,
-		Force:          true,
-	}
-
-	// 2. Install the Elastic-Agent with the policy that
-	// was just created.
-	policy, err := tools.InstallAgentWithPolicy(
-		ctx,
-		t,
-		installOpts,
-		agentFixture,
-		info.KibanaClient,
-		createPolicyReq)
-	require.NoError(t, err)
-	t.Logf("created policy: %s", policy.ID)
-	check.ConnectedToFleet(ctx, t, agentFixture, 5*time.Minute)
-
-	t.Run("Monitoring logs are shipped", func(t *testing.T) {
-		testMonitoringLogsAreShipped(t, ctx, info, agentFixture, policy)
-	})
-
-	t.Run("Normal logs with flattened data_stream are shipped", func(t *testing.T) {
-		testFlattenedDatastreamFleetPolicy(t, ctx, info, policy)
-	})
-}
-
 func startMockES(t *testing.T) string {
 	registry := metrics.NewRegistry()
 	uid := uuid.Must(uuid.NewV4())
